ai-agent/requirements.txt
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
comm==0.2.2
debugpy==1.8.2
decorator==5.1.1
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
executing==2.0.1
fastapi==0.111.1
fastapi-cli==0.0.4
fastembed==0.3.4
filelock==3.15.4
flatbuffers==24.3.25
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.64.1
grpcio-tools==1.64.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.5
humanfriendly==10.0
hyperframe==6.0.1
idna==3.7
ipykernel==6.29.5
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==4.1.0
mpmath==1.3.0
nest-asyncio==1.6.0
numpy==1.26.4
onnx==1.16.1
onnxruntime==1.18.1
openai==1.35.14
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
portalocker==2.10.1
prompt_toolkit==3.0.47
protobuf==5.27.2
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
PyStemmer==2.2.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
qdrant-client==1.10.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.19.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.5
stack-data==0.6.3
starlette==0.37.2
streamlit==1.36.0
sympy==1.13.0
tenacity==8.5.0
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typer==0.12.3
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uuid==1.30
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
wikipedia==1.4.0